fix: ensure `options` are passed to app plugin when `--options` flag is provided
The TS typings and docs strongly imply that the `options` object defined in the app.js
file should make their way to the app plugin if the `--options` cli argument is present
when running `fastify start app.js`. This fix does just that, providing the exported
options object back to the app plugin when it is registered.

The CLI allows arguments to make their way to the app via additional args (coming after
a `--`), eg: `fastify start app.js -l info -- --custom test` will populate the plugin
`options` argument with a `custom: "test"` property. This change deliberately leaves
precedence with the CLI args so that the exported `options` object just forms the default
values which can then be overridden as required.